[FORMAT] Reformat code using vcFormat with custom workspace settings to make it as close as possible to the current code
We also tried clang-format at first, and committed the best options we could find, but didn't choose it eventually as it was too far from the conventions we wanted. vcFormat is much more lax, which has the advantage of allowing us to format things as we want in some areas, but also will not detect bad formatting.

Most of the changes are in fact correct, as we forgot to properly format some parts of the code.
However, in some cases, unwanted changes were enforced due to known limitations:
- #endif will be indented when wrapping a debug-only variable

We also added custom changes for clarity:
- case should have wrapping braces, except if it only contains a break
- alphabetical order of includes in each section (except for Common.h at the top)